Closing issue #34: base layers broken on zoom
This was one terrible bug, and the results aren't very clean.

The problem was this.  The OpenLayers map object is supposed to have a
reference to the current baselayer object.  Unfortunately, the geoext
layer tree isn't setting it.  Huge refresh/zooming problems ensue.

The fix was to manually set map.baselayer on initialization, and
whenever the user clicks a radio button.  This got messy on
initialization, because I couldn't find a way to change the radio
button.  I eventually included jquery as a dependency and simulated a
mouse click.  Just terrible.

Methinks we wouldn't have had to do any of this if we knew how to use
geoext better.  I'm thinking that the way we're populating the layer
tree isn't right.  (We've had problems on this front in the past.)  It
seems like when it tries to build a base layer node it treats it as just
a simple overlay node.  Some data is lost in the transfer.

Also in this commit: fixing the opacity slider for real.  Now it works
with base layers, too.