Closed issue #51, aka Dave's issue.
Generally, now, if the user sees a dot then can click on it and get a
popup.

Additionally, the Edit and Legend panels are not inaccessible when the
user is viewing a base layer.  This was technically true before, but now
the panels are also grayed out.  Hopefully this makes it more obvious to
the user what's going on.